+++ Source/SelfService/Web/integrations/connections/ConnectionsDataGrid.tsx
@@ -5,26 +5,13 @@ import React from 'react';
 
 import { useNavigate } from 'react-router-dom';
 
-import { DataGridPro, GridColDef } from '@mui/x-data-grid-pro';
-import { Paper } from '@mui/material';
+import { DataGridPro } from '@mui/x-data-grid-pro';
 
-import { StatusIndicator } from '@dolittle/design-system';
+import { dataGridDefaultProps, DataGridWrapper } from '@dolittle/design-system';
 
 import { ConnectionModel } from '../../apis/integrations/generated';
-import { getIndicatorStatusFromStatusMessage } from '../statusHelpers';
 
-
-type ConnectionsDataGridRowProps = {
-    row: ConnectionModel;
-};
-
-const StatusCell = ({ row }: ConnectionsDataGridRowProps) => {
-    const status = getIndicatorStatusFromStatusMessage(row.status.statusMessage);
-
-    return (
-        <>{status && <StatusIndicator status={status.status} label={status.label} message={status.message} />}</>
-    );
-};
+import { connectionsDataGridColumns } from './ConnectionsDataGridColumns';
 
 export type ConnectionsDataGridProps = {
     connections: ConnectionModel[];
@@ -33,35 +20,6 @@ export type ConnectionsDataGridProps = {
 
 export const ConnectionsDataGrid = ({ connections, isLoading }: ConnectionsDataGridProps) => {
     const navigate = useNavigate();
-
-    const connectionsColumns: GridColDef<ConnectionModel>[] = [
-        {
-            field: 'name',
-            headerName: 'Name',
-            minWidth: 270,
-            flex: 1,
-        },
-        {
-            field: 'description',
-            headerName: 'Description',
-            minWidth: 270,
-            flex: 1,
-        },
-        {
-            field: 'source',
-            headerName: 'Source',
-            minWidth: 270,
-            flex: 1,
-            valueGetter: ({ row }: ConnectionsDataGridRowProps) => 'M3',
-        },
-        {
-            field: 'status',
-            headerName: 'Connection Status',
-            minWidth: 270,
-            flex: 1,
-            renderCell: StatusCell,
-        }
-    ];
 
     const handleRowClick = (connectionModel: ConnectionModel) => {
         if (connectionModel.connectionId) {
@@ -72,19 +30,14 @@ export const ConnectionsDataGrid = ({ connections, isLoading }: ConnectionsDataG
     };
 
     return (
-        <Paper sx={{ width: 1 }}>
+        <DataGridWrapper>
             <DataGridPro
+                {...dataGridDefaultProps}
                 rows={connections}
-                columns={connectionsColumns}
-                getRowHeight={() => 'auto'}
-                autoHeight
-                headerHeight={46}
-                disableColumnMenu
-                hideFooter
-                disableSelectionOnClick
+                columns={connectionsDataGridColumns}
                 loading={isLoading}
                 onRowClick={({ row }) => handleRowClick(row)}
             />
-        </Paper>
+        </DataGridWrapper>
     );
 };
